Don't allow player input if not playing.

+++ player.mjs
@@ -12,6 +12,7 @@ export class Player {
     this.level_ = 0;
     this.drop_timer_ = null;
     this.player_name_ = null;
+    this.is_playing_ = false;
 
     this.player_info_controller_ = player_info_controller;
     this.player_info_controller_.setStartGameCallback(
@@ -24,6 +25,7 @@ export class Player {
   }
 
   start(player_name) {
+    this.is_playing_ = true;
     this.player_name_ = player_name;
     this.newShape();
     this.reloadTimer();
@@ -38,6 +40,10 @@ export class Player {
   }
 
   drop() {
+    if (!this.is_playing_) {
+      return;
+    }
+
     if (!this.current_shape_.drop()) {
       this.newShape();
     }
@@ -46,21 +52,37 @@ export class Player {
   }
 
   rotate() {
+    if (!this.is_playing_) {
+      return;
+    }
+
     this.current_shape_.rotate();
     this.renderer_.draw();
   }
 
   left() {
+    if (!this.is_playing_) {
+      return;
+    }
+
     this.current_shape_.left();
     this.renderer_.draw();
   }
 
   right() {
+    if (!this.is_playing_) {
+      return;
+    }
+
     this.current_shape_.right();
     this.renderer_.draw();
   }
 
   dropAllTheWay() {
+    if (!this.is_playing_) {
+      return;
+    }
+
     while (this.current_shape_.drop()) {}
     this.newShape();
     this.renderer_.draw();
